Update homex.g
M574 command at the end of the homing script will overwrite the setting in config.g.
This is, because on my last commit, I requested to remove the initial M574 that did set up sensorless homing, but didn't saw the M574 command at the end, that also needed to be removed. Sorry.