Update bed.g
switching G29 on and off in the toolchange-macros triggers a warning-message "Warning: the height map was loaded when the current Z=0 datum was not determined probing. This may result in a height offset."
The simplest solution to this I could find is to add a "G30" right after the "G28 Z". 
This was suggested by DC42 at https://forum.duet3d.com/topic/10835/z-0-datum-was-not-determined-probing